# Weighting the samples by their alpha coverage (draft)