Hello,

first of all thank you for making this! I was recently thinking about the technique and it's nice to have a solid foundation to build upon.

One detail though does not seem to align with the steps outlined in the God of War talk -
>2. At each step we scale it down to half of its size, **weighting the samples by their alpha coverage.**

My interpretation is that this is to discard pixel values that are not inside the coverage mask.

A quick example why this is beneficial:
<p><img src="https://github.com/user-attachments/assets/4339989c-351a-4fd5-99ab-0051bae382fc" width=256 height=256> <img src="https://github.com/user-attachments/assets/becebada-d6ac-4264-849c-0657ddf4b184" width=256 height=256></p>


Compositing linearly scaled down mip levels (current implementation), we get this:
<img src="https://github.com/user-attachments/assets/43e3d3ee-55b3-48c3-b705-1f688b2d0689" width=256 height=256>

But that would potentially result in some darker tones bleeding into the shape, which is usually avoided using edge dilation:
<img src="https://github.com/user-attachments/assets/3c248edf-0ada-4494-96ef-2465154521b8" width=256 height=256>

Instead, the color values outside of coverage should be discarded when calculating the pixel average when scaling down.

I first tried an implementation using NumPy, where I did a value by value calculation, but that turned out to be too slow (best version was 2x the run time of the current implementation). What also works however is this approach:
1. Multiply color by alpha coverage, to zero out any color information outside of coverage
2. Then scale down color and coverage mask linearly (Image.reduce is perfect for that, Image.resize gave me issues)
3. Now re-normalize the color using the mask, essentially `color / mask`

The result:
<img src="https://github.com/user-attachments/assets/b0d42d3b-3098-4ff0-98dc-e9a570fbaee2" width=256 height=256>

This PR draft contains two variants
1. Using your method of directly compositing the layers
2. Using the method described in the God of War talk, which cumulatively processes and stores the mip levels and then composites them

The second variant has been ~significantly~ faster in my tests (which makes sense, as progressively less pixels have to be considered during scaling), but of course uses more RAM. The pure Pillow variant has some smaller issues with color precision, which I think is a result of cumulative rounding errors, since this is all done in uint8. The NumPy+Pillow variant is a little slower, but works with 32 bit precision.

Some numbers (a few 8 bit PNG images 1K to 8K, single thread):

1. Direct composition

| Method  | Time |
| :- | -: |
| Current | 44.82 s |
| Weighted | 40.32 s |

3. Cumulative

| Method  | Time |
| :- | -: |
| Walking Weighted | 38.90 s |
| Walking Weighted (NumPy) | 45.65 s |