@Level3Manatee 2024-08-15
Actually found a fast-ish way to do this with NumPy