@Level3Manatee 2024-08-15
Ok, the weighted adaption of the original implementation was unnecessarily slow, got it 25% faster by pre-multiplying only once, so that one looks pretty attractive now.

The NumPy version could be interesting for replacing Pillow with e.g. OpenCV or maybe SciKit, for processing images > 8 bit. Very possible that it could be faster still, haven't done too much with NumPy before.