@Level3Manatee 2024-08-17
Did some further testing and found incorrect results for a very dark texture with all but the NumPy variant ... I'm guessing rounding issues. Will look into it a bit more.
Numpy:
![correct](https://github.com/user-attachments/assets/cf3698cd-63d7-4608-8120-76d39c064e2c)
Others:
![incorrect](https://github.com/user-attachments/assets/f3a3d12d-9183-4eb1-918a-391c71b1a047)