@secarri 2024-08-17
Hello Jan! 

I was about to report this issue and had a chance to look into it yesterday. As we discussed, I’ve started porting the code to C#. However, I don’t expect significant speed improvements initially. For processing times under one second, C or C++ would be better. Python and C# are currently similar in performance. I’ve pushed my changes, though they’re not yet in the main branch. Everything is working, but some parts are still a bit slow. If you end up making a C++ version let me know :)

Your Fork             |  Expected, currently in the new branch
:-------------------------:|:-------------------------:
<img src="https://github.com/user-attachments/assets/f3de430d-f0ab-4048-aee1-c670e83c3f31" width="250" height="250" /> | <img src="https://github.com/user-attachments/assets/c4899ab0-e9d9-4608-bab4-fc9c0a23e057" width="250" height="250" />


New branch, still WIP, will push it to main early next week: [csharp-refactor-python-binding](https://github.com/secarri/MipFlooding/tree/csharp-refactor-python-binding)