@Level3Manatee 2024-08-17
Yeah essentially the issue is with the python-specific optimization route I took (pre-multiplying, using natively implemented scaling to calculate the average, then re-normalize), since with images that have little coverage, that is bound to run into precision issues at high resolutions (as averaging a mostly black image will yield very small numbers). Not really an issue with the walking approach though, especially at 32 bit precision (NumPy). Going to delete the other variants and fix the non-NumPy version and then call this one finished I guess. ;)