@Level3Manatee 2024-08-17
I decided to only keep the NumPy variant, as that one produces the most accurate results.

Output looks like this now:
<img src="https://github.com/user-attachments/assets/a26838d6-913d-4c40-8f33-face0abb8253" width=256 height=256>

The darkening was a result of the way masks were imported (mask is in R channel of an RGB image, converted to grayscale that makes a very dark mask).

Not sure when I'll get to a C++ implementation, but I've never written a Python<->Native integration so that definitely sounds interesting. ;) Thank you for the feedback, closing this now.